VirtualSeco/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.5
django-cors-headers==3.7.0
djangorestframework==3.12.2
future==0.18.2
idna==2.10
joblib==1.0.0
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.5
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
serializers==0.2.4
six==1.15.0
sqlparse==0.4.1
textblob==0.15.3
tqdm==4.56.0
tweepy==3.10.0
urllib3==1.26.3
gunicorn==20.0.4
